requirements.txt
pip
fastapi
bcrypt
PyJWT
passlib
deta
